+++ packages/components/components/elvis-header/react.js
@@ -1,1 +1,2 @@
 export { default as Header } from './dist/react/js/elvia-header.js';
+export { getCurrentTheme } from './dist/react/js/publicApi.js';