+++ packages/components/components/elvis-header/src/react/themePicker/themePicker.tsx
@@ -9,8 +9,9 @@ import {
 import { LightThemeIcon } from './lightThemeIcon';
 import { DarkThemeIcon } from './darkThemeIcon';
 import { SystemThemeIcon } from './systemThemeIcon';
-import { getStoredActiveTheme, setThemeClassOnDocument } from '../themeUtils';
-import { Theme, ThemeEvent, themeLocalStorageKey } from '../elviaHeader.types';
+import { getApplicableTheme, getStoredActiveTheme, setThemeClassOnDocument } from '../themeUtils';
+import { ThemeEvent } from '../elviaHeader.types';
+import { Theme, themeLocalStorageKey } from '../elviaHeader.shared.types';
 
 interface PickerTheme {
   theme: Theme;
@@ -33,7 +34,7 @@ export const ThemePicker: React.FC<ThemePickerProps> = ({ onThemeChange }) => {
     setThemeClassOnDocument(theme);
     setCurrentTheme(theme);
     localStorage.setItem(themeLocalStorageKey, theme);
-    onThemeChange?.(theme);
+    onThemeChange?.(getApplicableTheme(theme));
   };
 
   useEffect(() => {
@@ -42,6 +43,7 @@ export const ThemePicker: React.FC<ThemePickerProps> = ({ onThemeChange }) => {
     const onUserPreferenceChange = (change: MediaQueryListEvent): void => {
       if (currentTheme === 'system') {
         setThemeClassOnDocument(change.matches ? 'dark' : 'light');
+        onThemeChange?.(change.matches ? 'dark' : 'light');
       }
     };
     const prefersColorScheme = window.matchMedia('(prefers-color-scheme: dark)');