+++ packages/web/src/app/doc-pages/brand/icon-doc/icon-data.ts
@@ -10,10 +10,6 @@ const elvisIconData: ComponentData = {
       type: 'string',
       description: 'Available icons from elvia/elvis-assets-icons package, in camelCase',
       default: '""',
-      cegDisplayName: 'name',
-      cegDefault: 0,
-      cegType: 'string',
-      cegFormType: 'iconName',
     },
     color: {
       isRequired: false,
@@ -26,13 +22,6 @@ const elvisIconData: ComponentData = {
       type: `"xxs" ... =>  "xxl" | undefined`,
       description: `Standard size of icon, follows same sizing guide as css classes, e.g "xs" = 16px, "sm" = 24px `,
       default: '"sm"',
-      cegDisplayName: 'Size',
-      cegType: 'string',
-      cegDefault: 'sm',
-      cegFormType: 'radio',
-      cegOptions: ['xs', 'sm', 'md', 'lg', 'xl', 'xxl'],
-      cegOptionsLabel: ['XS', 'Small', 'Medium', 'Large', 'XL', 'XXL'],
-      cegLabelTypography: 'capitalize',
     },
     customSize: {
       isRequired: false,
@@ -41,28 +30,6 @@ const elvisIconData: ComponentData = {
       default: '',
     },
   },
-  codeReact: `<Icon
-  name={"addCircle"}
-></Icon>
-`,
-  codeAngular: `<elvia-icon
-  [name]="'addCircle'"
->
-</elvia-icon>
-`,
-  codeNativeHTML: `<elvia-icon
-  name="addCircle"
-  id="example-elvia-icon"
->
-</elvia-icon>
-`,
-  codeVue: `<elvia-icon
-  :name="'addCircle'"
->
-</elvia-icon>
-`,
-  codeNativeScript: `const elvisIcon = document.getElementById('example-elvia-icon');
-`,
 };
 
 export { elvisIconData };