+++ packages/web/src/app/doc-pages/brand/typography-doc/typography-doc.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
 import { TypographyDocComponent } from './typography-doc.component';
 import { ComponentHeaderModule } from 'src/app/shared/component-documentation/component-structure/component-header/component-header.module';
 import { ComponentSubsubsectionModule } from 'src/app/shared/component-documentation/component-structure/component-subsubsection/component-subsubsection.module';
@@ -23,6 +24,7 @@ import '@elvia/elvis-tabs';
     ComponentSubsectionModule,
     ComponentSubsubsectionModule,
     ComponentChangelogModule,
+    RouterModule,
   ],
   declarations: [
     TypographyDocComponent,