+++ packages/web/src/app/shell/header/header.component.ts
@@ -123,6 +123,7 @@ export class HeaderComponent {
     const classToRemove: ThemeClassName = theme === 'light' ? 'e-theme-dark' : 'e-theme-light';
     const classToAdd: ThemeClassName = theme === 'light' ? 'e-theme-light' : 'e-theme-dark';
 
+    document.body.style.colorScheme = theme;
     document.body.classList.remove(classToRemove);
     if (!document.body.classList.contains(classToAdd)) {
       document.body.classList.add(classToAdd);